+++ packages/nimble-components/src/text-field/types.ts
@@ -1,4 +1,25 @@
 import { TextFieldType } from '@microsoft/fast-foundation';
 
-export { TextFieldType };
+/**
+ * Values for the 'type' attribute of the text field.
+ */
 export type TextFieldTypeAttribute = `${TextFieldType}`;
+
+/**
+ * Values for the 'type' property of the text field.
+ */
+export { TextFieldType };
+
+/**
+ * Values for the 'appearance' attribute of the text field.
+ */
+export type TextFieldAppearanceAttribute = `${TextFieldAppearance}`;
+
+/**
+ * Values for the 'appearance' property of the text field
+ */
+export enum TextFieldAppearance {
+    Underline = 'underline',
+    Outline = 'outline',
+    Block = 'block'
+}