+++ angular-workspace/projects/ni/nimble-angular/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Fri, 11 Feb 2022 16:48:53 GMT and should not be manually modified.
+This log was last generated on Tue, 15 Feb 2022 17:32:39 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.1.4
+
+Tue, 15 Feb 2022 17:32:39 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v2.0.0
+
+## 1.1.3
+
+Mon, 14 Feb 2022 21:34:16 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v1.2.0
+
+## 1.1.2
+
+Mon, 14 Feb 2022 18:58:21 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v1.1.2
 
 ## 1.1.1